Fix compiler warnings

+++ examples/FeatureShowcase/Source/AnimationGraphTest/AnimationGraphTest.cpp
@@ -274,8 +274,8 @@ void AnimationGraphTest::Draw()
     commandList->SetRenderPass(std::move(pass));
 
     auto projectionMatrix = Matrix4x4::CreateOrthographicLH(
-        presentationParameters.BackBufferWidth,
-        presentationParameters.BackBufferHeight,
+        static_cast<float>(presentationParameters.BackBufferWidth),
+        static_cast<float>(presentationParameters.BackBufferHeight),
         0.0f,
         100.0f);
